remove testing code
