+++ app/server.js
@@ -18,7 +18,6 @@ import { requestIdMiddleware } from 'api/utils/requestIdMiddleware';
 import { Redis } from 'api/infrastructure/Redis';
 import { maskMongoPassword } from 'api/utils/maskMongoPassword';
 import { elasticClient } from 'api/search/elastic';
-import 'api/core/infrastructure/listeners/Listeners';
 import uwaziMessage from '../message';
 import apiRoutes from './api/api';
 import privateInstanceMiddleware from './api/auth/privateInstanceMiddleware';
